Solutions more lessons

+++ 1.JavaSyntax/src/com/javarush/task/pro/task07/task0709/Solution.java
@@ -6,7 +6,7 @@ package com.javarush.task.pro.task07.task0709;
 
 public class Solution {
     long a = 109 + 15;
-    int b = a * 2;
-    short c = a / b;
-    byte d = a + b - c;
+    int b = (int)a * 2;
+    short c = (short)(a / b);
+    byte d = (byte)(a + b - c);
 }
